# Make 'Terminal.renderLazy' lazy