--- a/prettyprinter-ansi-terminal/prettyprinter-ansi-terminal.cabal
+++ b/prettyprinter-ansi-terminal/prettyprinter-ansi-terminal.cabal
@@ -67,23 +67,16 @@ benchmark large-output
     build-depends:
           base >= 4.5 && < 5
         , base-compat >=0.9.3 && <0.12
+        , containers
+        , deepseq
+        , gauge >= 0.2
         , prettyprinter
         , prettyprinter-ansi-terminal
-
-        , gauge >= 0.2
         , QuickCheck >= 2.7
-        , containers
         , text
-        , deepseq
 
     hs-source-dirs:      bench
     main-is:             LargeOutput.hs
-    ghc-options:         -threaded -rtsopts -with-rtsopts=-N -Wall
+    ghc-options:         -O2 -threaded -rtsopts -with-rtsopts=-N -Wall
     default-language:    Haskell2010
     type:                exitcode-stdio-1.0
-
-    if !impl(ghc >= 7.6)
-        build-depends: ghc-prim
-
-    if !impl(ghc >= 8.0)
-        build-depends: semigroups