@sjakobi 2020-07-01
> Any suggestions on how to benchmark this?

Good question! I guess a somewhat lasting solution would be to port the [`large-output` benchmark](https://github.com/quchen/prettyprinter/blob/master/prettyprinter/bench/LargeOutput.hs) from `prettyprinter` and add some annotations. No need to include all the different layouters though. Just using the fastest one should be enough. That's probably `layoutCompact`.

Does that sound reasonable?