@georgefst 2020-07-02
> Does that sound reasonable?

Certainly does. I probably won't get around to it until next week though.