@georgefst 2020-07-12
Well... We actually seem to be looking at a more than 2x speedup.

Although, the benchmark data doesn't currently actually contain any annotations. I may add some and see what happens.