@sjakobi 2020-07-12
> Well... We actually seem to be looking at a more than 2x speedup.

Wow! Any idea why?

> Although, the benchmark data doesn't currently actually contain any annotations. I may add some and see what happens.

Yes, please do!