@georgefst 2020-07-12
> Wow! Any idea why?

Not really. All I'd say is that the old implementation is in `ST`, seemingly in order to mirror the implementation of `renderIO`, which maybe introduces a significant overhead. Not sure as I've never used it, but it's certainly not the obvious approach.