@georgefst 2020-07-13
I've played around with this a bit, and adding annotations only amplifies the difference. As does using `renderStrict`, for some reason, even though it's still just `toStrict . renderLazy`.

The question now is which benchmarks are worth keeping around. We've got a lot of choices:
- `renderStrict`, `renderLazy`
- comparison with `ansi-wl-pprint`
- light annotations, heavy annotations, none

And using annotations also means using an algorithm other than `layoutCompact` - I'm not sure which would be best. (Side note - it's really not obvious from the docs that `layoutCompact` strips annotations.)