@sjakobi 2020-07-13
> I've played around with this a bit, and adding annotations only amplifies the difference. As does using `renderStrict`, for some reason, even though it's still just `toStrict . renderLazy`.

Interesting. Have you looked at the Core to get more insight into the reason for the performance difference?


> The question now is which benchmarks are worth keeping around. We've got a lot of choices:
> 
> * `renderStrict`, `renderLazy`

`renderLazy` should be sufficient IMHO

> * comparison with `ansi-wl-pprint`

I don't care too much about that, although it's a nice baseline.

It might be nice to include `prettyprinter`'s `renderLazy` as another baseline though.

> * light annotations, heavy annotations, none

Heavy annotations seem most interesting to me.

> And using annotations also means using an algorithm other than `layoutCompact` - I'm not sure which would be best. (Side note - it's really not obvious from the docs that `layoutCompact` strips annotations.)

Oops! I didn't know that. Let's use `layoutPretty` then.