@georgefst 2020-07-13
I've updated the benchmark with `layoutPretty defaultLayoutOptions`, and lots of annotations, using `Text.renderLazy` as the baseline.

> Have you looked at the Core to get more insight into the reason for the performance difference?

Nope. I don't have any experience with reading Core, and personally I'm happy to just defer to the magical black box that is GHC on this one.