@sjakobi 2020-07-13
Nice work, @georgefst! Could you paste the benchmark results with the old and new implementations here?

Regarding correctness: I'm unsure how much we can rely on the few doctests in `prettyprinter-ansi-terminal`. Have you tested this code in other contexts? Would the `pretty-simple` tests be robust enough to detect any issues?