@georgefst 2020-07-13
@sjakobi Well, there's:
- the tests here
- the tests in `pretty-simple`
- all the examples I used while benchmarking
- the fact that it was adapted quite directly from the old implementation anyway

So I'm *reasonably* confident.