@georgefst 2020-07-13
Old implementation:
```
benchmarking prettyprinter-ansi-terminal ... took 151.4 s, total 56 iterations
benchmarked prettyprinter-ansi-terminal
time                 2.827 s    (2.543 s .. 3.045 s)
                     0.988 R²   (0.972 R² .. 0.998 R²)
mean                 2.663 s    (2.499 s .. 2.790 s)
std dev              281.3 ms   (205.9 ms .. 378.3 ms)
variance introduced by outliers: 29% (moderately inflated)
```

New version:
```
benchmarking prettyprinter-ansi-terminal ... took 65.21 s, total 56 iterations
benchmarked prettyprinter-ansi-terminal
time                 1.242 s    (1.028 s .. 1.558 s)
                     0.941 R²   (0.892 R² .. 0.991 R²)
mean                 1.202 s    (1.112 s .. 1.326 s)
std dev              182.8 ms   (114.6 ms .. 268.5 ms)
variance introduced by outliers: 48% (moderately inflated)
```