@georgefst 2020-07-14
@sjakobi The review comments I haven't replied to all relate to things that were copy-pasted from the benchmark from the core library. I'm happy to implement all your suggestions, unless you can see any reason not to.

Presumably any backwards compatibility issues would be caught by CI?