@georgefst 2020-07-14
> Presumably any backwards compatibility issues would be caught by CI?

By which I mean, specifically, compiling with old GHCs.