@sjakobi 2020-07-14
Yeah, we build the benchmarks in a few CI jobs, and I don't really care about the compatibility of the benchmarks with older GHC versions. So if you could clean up these wibbles a bit, that would be appreciated.

---

@quchen Modulo the mentioned wibbles, this looks ready to go to me. Would you like to review, too?