@sjakobi 2020-07-20
What's the status now, @georgefst? Is this ready?! :)

Have the benchmark results changed?