@georgefst 2020-07-20
> What's the status now, @georgefst? Is this ready?! :)

Yes it's ready, unless we care about the one remaining review comment. I'd be happy to leave it as-is, seeing as it's no worse than the existing benchmark in the core library.

> Have the benchmark results changed?

Slight speedup with `-O2` - around 20% for the new implementation, about 5% with the old one.