@georgefst 2020-07-20
> wibble

I can't find any actual definition of it, but somehow I instantly knew what you meant. And I can't think of a good alternative, although "quibble" and "foible" are somewhat related. English is a silly language, and I've noticed SPJ is a fairly creative user of it.

Anyway, it amused me because it instantly reminded me of [a scene from a classic TV show](https://www.youtube.com/watch?v=WFZrseGNlC8), which is, as it happens, what almost all references to "wibble" on the web seem to refer to.