@sjakobi 2020-07-21
FYI regarding the release: https://github.com/quchen/prettyprinter/issues/181